Fix: 관리자 회원가입, 로그인, 검색, 수정, 삭제 API 구현

+++ api/admin/service.js
@@ -2,15 +2,15 @@ const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 const Admin = require('./model');
 const { Op } = require('sequelize');
+const { getPagination, getPagingData } = require('../../utils');
+const { BadRequestClass, NotFoundClass } = require('../../middlewares');
 
 module.exports = {
   async addAdmin(adminInfo) {
     const { name, email, password } = adminInfo;
     const admin = await Admin.findOne({ where: { email: email } });
     if (admin) {
-      throw new Error(
-        '이 이메일은 현재 사용중입니다. 다른 이메일을 입력해 주세요.'
-      );
+      throw new BadRequestClass('이미 존재하는 이메일입니다.');
     }
 
     const hashedPassword = await bcrypt.hash(password, 10);
@@ -24,14 +24,12 @@ module.exports = {
     const { email, password } = adminInfo;
     const admin = await Admin.findOne({ where: { email: email } });
     if (!admin) {
-      throw new Error('이메일 정보가 맞지 않습니다. 다시 한 번 확인해 주세요.');
+      throw new NotFoundClass('이메일 입력값이 올바르지 않습니다.');
     }
 
     const isPasswordCorrect = await bcrypt.compare(password, admin.password);
     if (!isPasswordCorrect) {
-      throw new Error(
-        '비밀번호 정보가 맞지 않습니다. 다시 한 번 확인해 주세요.'
-      );
+      throw new NotFoundClass('비밀번호 입력값이 올바르지 않습니다.');
     }
 
     const secretKey = process.env.JWT_SECRET_KEY;
@@ -40,37 +38,48 @@ module.exports = {
     return token;
   },
 
-  async getAdmin(id) {
-    const admin = await Admin.findOne({ where: { id: id } });
-    if (!admin) {
-      throw new Error('가입 내역이 없습니다. 다시 한 번 확인해 주세요.');
-    }
-
-    return admin;
-  },
-
-  async getAdmins(adminInfo) {
-    const admins = await Admin.findAll({
-      where: {
-        [Op.or]: [
-          { name: { [Op.like]: `%${adminInfo}%` } },
-          { email: { [Op.like]: `%${adminInfo}%` } },
+  async getAdmins(page, size, keyword) {
+    const { limit, offset } = getPagination(page, size);
+    let admins;
+    if (!keyword || keyword === ':keyword') {
+      admins = await Admin.findAndCountAll({
+        order: [
+          ['name', 'ASC'],
+          ['email', 'ASC'],
         ],
-      },
-    });
+        offset,
+        limit,
+      });
+    } else {
+      admins = await Admin.findAndCountAll({
+        where: {
+          [Op.or]: [
+            { name: { [Op.like]: `%${keyword}%` } },
+            { email: { [Op.like]: `%${keyword}%` } },
+          ],
+        },
+        order: [
+          ['name', 'ASC'],
+          ['email', 'ASC'],
+        ],
+        offset,
+        limit,
+      });
+    }
+    admins = getPagingData(admins, page, limit);
+
     return admins;
   },
 
   async setAdmin(email, toUpdate) {
     const { name, password } = toUpdate;
-    let admin = await Admin.findAll({ where: { email: email } });
+    const admin = await Admin.findOne({ where: { email: email } });
     if (!admin) {
-      throw new Error('가입 내역이 없습니다. 다시 한 번 확인해 주세요.');
+      throw new NotFoundClass('존재하지 않는 이메일입니다.');
     }
 
     const newPasswordHash = await bcrypt.hash(password, 10);
-
-    admin = Admin.update(
+    const updateCount = Admin.update(
       {
         name: name,
         password: newPasswordHash,
@@ -79,14 +88,26 @@ module.exports = {
         where: { email: email },
       }
     );
+    if (updateCount < 1) {
+      throw new BadRequestClass(
+        `${name} 관리자 정보 수정 처리에 실패했습니다.`
+      );
+    }
 
-    return admin;
+    return updateCount;
   },
 
   async deleteAdmin(email) {
+    const admin = await Admin.findOne({ where: { email: email } });
+    if (!admin) {
+      throw new NotFoundClass('존재하지 않는 관리자 이메일입니다.');
+    }
+
     const deleteCount = await Admin.destroy({ where: { email: email } });
     if (deleteCount < 1) {
-      throw new Error(`${email} 관리자 탈퇴 처리에 실패하였습니다.`);
+      throw new BadRequestClass(
+        `${email} 관리자 정보 삭제 처리에 실패했습니다.`
+      );
     }
     return deleteCount;
   },
